dali/quick-start/requirements.txt
--extra-index-url https://developer.download.nvidia.com/compute/redist
nvidia-dali-cuda110
numpy
matplotlib
